+++ scripts/benchmark.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-asv run --show-stderr >asv-run.log
-asv publish