+++ scripts/run_tests.sh
@@ -3,6 +3,5 @@
 #   $ ./scripts/run_tests.sh
 # or
 #   $ ./scripts/run_tests.sh --cov pycvodes --cov-report html
-python2.7 -m pytest
-python3 -m pytest --doctest-modules --pep8 --flakes $@
-python3 -m doctest README.rst
+${PYTHON:-python} -m pytest --doctest-modules --pep8 --flakes $@
+${PYTHON:-python} -m doctest README.rst